fix: change type of parameters to adhere to specification of EBFilter

+++ snappy_pipeline/workflows/somatic_variant_filtration/model.py
@@ -31,16 +31,16 @@ class FilterSets(SnappyModel):
 class EbfilterSet(SnappyModel):
     shuffle_seed: int = 1
     panel_of_normals_size: int = 25
-    min_mapq: float = 20
-    min_baseq: float = 15
+    min_mapq: int = 20
+    min_baseq: int = 15
 
 
 class Ebfilter(SnappyModel):
     ebfilter_threshold: float = 2.4
     shuffle_seed: int = 1
     panel_of_normals_size: int = 25
-    min_mapq: float = 20
-    min_baseq: float = 15
+    min_mapq: int = 20
+    min_baseq: int = 15
 
 
 class Dkfz(SnappyModel):
